refactor styles loc and server ts
