+++ src/components/Dice/Die.js
@@ -5,7 +5,7 @@
 //Imports
 import React, { useEffect, useState , useRef } from 'react';
 import PropTypes from 'prop-types';
-import classes from './Die.module.css';
+import classes from '../../styles/Die.module.css';
 import Card from '@material-ui/core/Card';
 import DiceModel from './DiceModel';
