+++ src/styles/Die.module.css
